+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Version 0.6.3 (TBA)
+* New Feature: Added the connection feature manager
+* New Feature: Added sequence number audit (BitfinexConnectionFeature.SEQ_ALL)
 * Bugfix: Added missing BigDecimal order constructor (closes #23)
 
 # Version 0.6.2 (06.04.2018)